backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.12.0
sentence-transformers==2.3.1
faiss-cpu==1.8.0
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
google-auth==2.27.0
requests==2.31.0
numpy>=1.26.0
pymongo==4.6.1
PyJWT==2.8.0
bcrypt==4.1.2

python-pptx==0.6.23
